+++ tfimm/layers/__init__.py
@@ -3,7 +3,7 @@ from .blurpool import BlurPool2D  # noqa: F401
 from .classifier import ClassifierHead  # noqa: F401
 from .drop import DropPath  # noqa: F401
 from .factory import act_layer_factory, norm_layer_factory  # noqa: F401
-from .std_conv import StdConv2D  # noqa: F401
+from .conv import PadConv2D, PadDepthwiseConv2D, StdConv2D  # noqa: F401
 from .transformers import (  # noqa:F401
     MLP,
     ConvMLP,
